Tried to test flat and bind layout issues
`interpolate(flat)` is problematic in OpenGL because OpenGL uses
the last vertex of each triangle for the value passed to the fragment
shader where as WebGPU uses the first

So, added code to print out a warning if the shader uses
`interpolate(flat)`. So far, the uses I found would work with
either provoking-vertex = first or last. Common usage include
passing @builtin(instance_index) to the fragment shader (three.js)
and equivalent (cornell box on webgpu samples) as well as for
some reason, babylon.js passing a texture size from vertex
to fragment shader (🤷‍

The bind layout issue is the fact that bind group layouts
from pipeline's created with `layout: 'auto'` are supposed
to be incompatible with any other bind group layout from
a different pipeline. Chrome is not enforcing this rule
correctly and so it's possible pages are accidentally using
this feature. I couldn't find an example of any using this
method though I only checked three.js, babylon.js, unity's
boat race, compute.toys, and webgpu-samples

One thing that was very frustrating is that for some reason,
the extension injects loading .js files by inserting
`<script>` tags but they load in a random order. My workaround
was to load the script I need loaded first 10 times followed
by the one I need last. That seems broken to me. AFAIK scripts
are supposed to execute in order.