Fix settings used in content scripts.
Pre 0.0.24 the settings were stored in the page's
session storage. This had the plus that the page could
read the settings synchronously and act on them, in particular
it was uses to inject scripts.

0.0.24 switched to the settings being stored in the extension's
session storage and dynamically injecting content scripts. That
worked but it broke the scripts that needed more than just on/off
(injected vs not injected), but they needed some actual settings.
Those were block-features, breakpoints, and raf-skip-frames.

You'd think that would be easy but the content scripts have to
be injected as world: "MAIN" and as such are not allowed to call
the extension.

The workaround is to inject a helper that runs in world: "ISOLATED"
that *can* talk to the extension. So, it relies a message from
the content-scripts to the service-worker and back. PITA!

Anyway it seems to be working.