Fix breakpoints
The issue was that breakpoint wrappers were not added until
the settings message arrived. So, even though we prevented
the user from getting an adapter until we got the settings,
they may have already extracted and cached the method functions.
So, instead, we wrap the entire API immediately and use a Set
to decide which functions to break on. This way they are wrapped
before the page gets a chance to cache them.